refactor(mcp): decompose mcp.service.ts into per-scope tool modules
mcp.service.ts was 908 raw lines — lint passed only because the giant
tool-description string literals count as comments under the max-lines
skipComments rule, so the next tool addition would trip the 800 gate.
Split the registrations into sibling modules following the proven
community-tools.ts pattern (registerXxxTools(server, companyId, deps)
free functions):

- read-tools.ts    — search + entity-read tools + read resources +
                     buildProgressReporter (ReadToolDeps)
- procedural-tools.ts — match_procedure / list_procedures
- write-tools.ts   — record/link/retract/record_procedure/retire +
                     registerAdminTools (forget_entity)

mcp.service.ts (908 → 142 lines) keeps the DI seam, HEALTH_TOOLS, the
health probe, and the scope-gated wiring in buildServer; the constructor
signature is unchanged. Pure refactor — same registered surface, same
scope gates: the mcp-tools.unit-spec scope/name assertions and 128 e2e
stay green. Adding a tool now touches a focused ~200-line module, not an
800-line wall.
